New feature non-flight

+++ _components/serviceList/expansionForm.vue
@@ -1,5 +1,5 @@
 <script lang="ts">
-import {defineComponent, computed, ref} from 'vue';
+import Vue, {defineComponent, computed, ref, getCurrentInstance} from 'vue';
 import serviceListStore from './store/serviceList';
 import postFavourites from './services/postFavourites'
 import deleteFavourites from './services/deleteFavourites'
@@ -17,18 +17,27 @@ export default defineComponent({
   setup(props) {
     const data: any = computed(() => props.data);
     const isDesktop = computed(() => (window as any).innerWidth >= '900');
+    const proxy = (getCurrentInstance() as any).proxy as any;
+    const permissionFavourite: any = computed(() => ({
+      create: Vue.prototype.$auth.hasAccess('isite.favourites.create'),
+      edit: Vue.prototype.$auth.hasAccess('isite.favourites.edit'),
+      index: Vue.prototype.$auth.hasAccess(`isite.favourites.index`),
+      destroy: Vue.prototype.$auth.hasAccess(`isite.favourites.destroy`),
+    }));
     async function selectFavourite(data: any) {
       data.favourite = !data.favourite;
       if (!data.favourite) {
+        if(!permissionFavourite.value.destroy) return;
         const favoriteData = serviceListStore().getFavouriteList().find(item => item.id === data.id);
         await deleteFavourites(favoriteData);
         serviceListStore().removeFromFavouriteList(data);
         alert.success({message: `Favorite deleted successfully Product:${data.title}`} );
       } else {
-        const response: any = await postFavourites(data);
+        if(!permissionFavourite.value.create) return;
+        const response: any = await postFavourites({...data, userId: proxy.$root.$store.state.quserAuth.userId});
         const favoriteData = {...data, favouriteId: response?.id} 
         serviceListStore().pustFavouriteList(favoriteData);
-        alert.success(`Favorite created successfully Product:${data.title}`); 
+        alert.success(`Favorite created successfully Product:${data.title}`);
       }
       await workOrderList().getFavourites(true);
     }
@@ -46,7 +55,8 @@ export default defineComponent({
       data, 
       favourite,
       selectFavourite,
-      refData
+      refData,
+      permissionFavourite
     }
   },
 })
@@ -57,7 +67,7 @@ export default defineComponent({
       <q-list v-for="(item, index) in data" :key="index">
         <q-expansion-item header-class="text-white">
           <template v-slot:header>
-            <q-item-section avatar class="q-pr-none " style="min-width: 45px;">
+            <q-item-section v-if="permissionFavourite.create" avatar class="q-pr-none " style="min-width: 45px;">
               <i 
                   class="fa-star color-icon-star tw-cursor-pointer tw-text-2xl"
                   @click="selectFavourite(data[index])"
@@ -68,13 +78,13 @@ export default defineComponent({
               />
             </q-item-section>
             <q-item-section class="q-py-sm">
-              <span class="tw-text-base tw-font-bold" style="color:#1F294F;">
+              <p class="tw-text-base tw-font-bold" style="color:#1F294F;">
                 {{ item.title }}
                 <br>
                 <span v-if="item.helpText" class="tw-text-xs tw-text-gray-500">
                   {{ item.helpText }}
                 </span>
-              </span>
+              </p>
               <span class="tw-text-sm" style="color:#8A98C3;">{{ showValue(item.formField.quantity) }}</span>
             </q-item-section>
           </template>
@@ -110,14 +120,14 @@ export default defineComponent({
         >
         <div 
           class="
-            tw-flex 
-            tw-w-2/5 
-            tw-truncate 
+            tw-flex
+            tw-w-2/5
+            tw-break-words
             tw-py-3 
             text-services 
             tw-pl-2"
           >
-              <div class="q-px-sm">
+              <div class="q-px-sm" v-if="permissionFavourite.create">
                 <i 
                   class="fa-star color-icon-star tw-cursor-pointer"
                   @click="selectFavourite(data[index])"
@@ -128,13 +138,13 @@ export default defineComponent({
                 />
               </div>
               <div>
-                <p>{{ item.title }} 
-                  <br>
-                  <span 
-                    v-if="item.helpText" 
+                <p>
+                  {{ item.title }}
+                </p>
+                <p
+                    v-if="item.helpText"
                     class="tw-text-xs tw-text-gray-500">
-                    {{ item.helpText }}
-                  </span>
+                  {{ item.helpText }}
                 </p>
               </div>
         </div>
@@ -142,7 +152,7 @@ export default defineComponent({
           class="
             tw-w-3/5 
             tw-mx-2 
-            tw-truncate 
+            tw-truncate
             tw-flex 
             tw-flex-wrap 
             tw-justify-end tw-gap-4"
